inline all trading logic in single example
