+++ README.md
@@ -1,8 +1,8 @@
 # Trading System - PECES
 
-This repository contains an early implementation of the **PECES** module.
-Each `Pez` represents a trading agent that operates on different
-market timeframes:
+This repository contains an early implementation of the **PECES** module
+condensed into a single script for simplicity.  Each `Pez` represents a
+trading agent that operates on different market timeframes:
 
 - **Sardina** – 1 hour candles
 - **Atun** – daily candles (executed at 4pm, Monday to Saturday)
@@ -16,8 +16,17 @@ the price returns to the open, a position is opened in the opposite
 direction.
 
 Positions include a take profit and stop loss based on the distance to
-the tail. If no exit condition is met before the period ends, the
+the tail. Once price moves halfway from the entry to the take profit
+level, the stop loss is moved from the tail to a new level just 10% of
+that distance away from the entry to lock in a small gain. These rules
+apply indistintamente para operaciones de **compra** (largas) o **venta**
+(cortas). If no exit condition is met before the period ends, the
 position is closed at the current market price.
 
 Trade results can be stored in `trade_log` and later exported to an
 Excel file using `pandas`.
+
+All of the classes and a small demonstration live inside the single
+`example.py` file.  Running it will open a long trade and then a short
+trade using the same `Sardina` instance, showing that the rules work
+igual para ambos sentidos.
